requirements.txt
slackclient
loguru